# backend/app/api/routes.py
"""API routes — generated from the APS OpenAPI design (docs/openapi.json).

Each endpoint's method + path + params are real; the body is a TODO stub.
Generated by Autonomous Product Studio (APS) - fill in the TODOs to ship.
"""
from __future__ import annotations

from fastapi import APIRouter, HTTPException

router = APIRouter()


@router.get("/users")
def get_users():
    """List Users."""
    # TODO: implement GET /users
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/users")
def post_users():
    """Create User."""
    # TODO: implement POST /users
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/users/{id}")
def get_users_x(id: str):
    """Get User."""
    # TODO: implement GET /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/users/{id}")
def put_users(id: str):
    """Update User."""
    # TODO: implement PUT /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/users/{id}")
def delete_users(id: str):
    """Delete User."""
    # TODO: implement DELETE /users/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/quizes")
def get_quizes():
    """List Quizes."""
    # TODO: implement GET /quizes
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/quizes")
def post_quizes():
    """Create Quiz."""
    # TODO: implement POST /quizes
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/quizes/{id}")
def get_quizes_x(id: str):
    """Get Quiz."""
    # TODO: implement GET /quizes/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/quizes/{id}")
def put_quizes(id: str):
    """Update Quiz."""
    # TODO: implement PUT /quizes/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/quizes/{id}")
def delete_quizes(id: str):
    """Delete Quiz."""
    # TODO: implement DELETE /quizes/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/plateforms")
def get_plateforms():
    """List Plateforms."""
    # TODO: implement GET /plateforms
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/plateforms")
def post_plateforms():
    """Create Plateform."""
    # TODO: implement POST /plateforms
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/plateforms/{id}")
def get_plateforms_x(id: str):
    """Get Plateform."""
    # TODO: implement GET /plateforms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/plateforms/{id}")
def put_plateforms(id: str):
    """Update Plateform."""
    # TODO: implement PUT /plateforms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/plateforms/{id}")
def delete_plateforms(id: str):
    """Delete Plateform."""
    # TODO: implement DELETE /plateforms/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/differentiators")
def get_differentiators():
    """List Differentiators."""
    # TODO: implement GET /differentiators
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/differentiators")
def post_differentiators():
    """Create Differentiator."""
    # TODO: implement POST /differentiators
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/differentiators/{id}")
def get_differentiators_x(id: str):
    """Get Differentiator."""
    # TODO: implement GET /differentiators/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/differentiators/{id}")
def put_differentiators(id: str):
    """Update Differentiator."""
    # TODO: implement PUT /differentiators/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/differentiators/{id}")
def delete_differentiators(id: str):
    """Delete Differentiator."""
    # TODO: implement DELETE /differentiators/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/statefuls")
def get_statefuls():
    """List Statefuls."""
    # TODO: implement GET /statefuls
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/statefuls")
def post_statefuls():
    """Create Stateful."""
    # TODO: implement POST /statefuls
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/statefuls/{id}")
def get_statefuls_x(id: str):
    """Get Stateful."""
    # TODO: implement GET /statefuls/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/statefuls/{id}")
def put_statefuls(id: str):
    """Update Stateful."""
    # TODO: implement PUT /statefuls/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/statefuls/{id}")
def delete_statefuls(id: str):
    """Delete Stateful."""
    # TODO: implement DELETE /statefuls/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/memories")
def get_memories():
    """List Memories."""
    # TODO: implement GET /memories
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.post("/memories")
def post_memories():
    """Create Memory."""
    # TODO: implement POST /memories
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.get("/memories/{id}")
def get_memories_x(id: str):
    """Get Memory."""
    # TODO: implement GET /memories/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.put("/memories/{id}")
def put_memories(id: str):
    """Update Memory."""
    # TODO: implement PUT /memories/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")

@router.delete("/memories/{id}")
def delete_memories(id: str):
    """Delete Memory."""
    # TODO: implement DELETE /memories/{id}
    raise HTTPException(status_code=501, detail="Not implemented yet")