# backend/app/models/schema.py
"""Database models — generated from the APS Technical Design data model.

Generated by Autonomous Product Studio (APS) - fill in the TODOs to ship.
"""
from __future__ import annotations

import uuid

from sqlalchemy import (Column, String, Integer, BigInteger, Float, Boolean,
                        DateTime, Date, Text, JSON, ForeignKey)  # noqa: F401
from sqlalchemy.dialects.postgresql import UUID
from sqlalchemy.orm import declarative_base

Base = declarative_base()


class User(Base):
    """User."""
    __tablename__ = "users"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    email = Column(String(255))
    role = Column(String(255))
    created_at = Column(DateTime(timezone=True))


class Quiz(Base):
    """Quiz."""
    __tablename__ = "quizs"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    owner_id = Column(UUID(as_uuid=True))  # TODO: foreign key — point to the referenced table
    name = Column(String(255))
    status = Column(String(255))
    created_at = Column(DateTime(timezone=True))
    updated_at = Column(DateTime(timezone=True))


class Plateform(Base):
    """Plateform."""
    __tablename__ = "plateforms"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    owner_id = Column(UUID(as_uuid=True))  # TODO: foreign key — point to the referenced table
    name = Column(String(255))
    status = Column(String(255))
    created_at = Column(DateTime(timezone=True))
    updated_at = Column(DateTime(timezone=True))


class Differentiator(Base):
    """Differentiator."""
    __tablename__ = "differentiators"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    owner_id = Column(UUID(as_uuid=True))  # TODO: foreign key — point to the referenced table
    name = Column(String(255))
    status = Column(String(255))
    created_at = Column(DateTime(timezone=True))
    updated_at = Column(DateTime(timezone=True))


class Stateful(Base):
    """Stateful."""
    __tablename__ = "statefuls"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    owner_id = Column(UUID(as_uuid=True))  # TODO: foreign key — point to the referenced table
    name = Column(String(255))
    status = Column(String(255))
    created_at = Column(DateTime(timezone=True))
    updated_at = Column(DateTime(timezone=True))


class Memory(Base):
    """Memory."""
    __tablename__ = "memorys"
    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    owner_id = Column(UUID(as_uuid=True))  # TODO: foreign key — point to the referenced table
    name = Column(String(255))
    status = Column(String(255))
    created_at = Column(DateTime(timezone=True))
    updated_at = Column(DateTime(timezone=True))